feat(api): extract Matching read model from discover
discover inlined 6 queries + grouping + scorer staging + pagination in a 750-LOC router with no read seam. Move the query side to matching-read-model.ts (getRankedCandidates), mirroring meetup-read-model. scoreCandidates stays in matching-utils. matching.ts 750->609. +4 harness tests.
